Add files via upload
A Calculator Project with basic arithmetic operations functionalities like addition, subtraction, multiplication, and division will be projects highly useful to a novice who has the essential knowledge of HTML, CSS, and JavaScript. You'll need to design a user interface with buttons for entering data and a display screen for displaying the results.